// nest/src/common/aligo/interfaces/aligo-template.interface.ts
export interface AligoTemplate {
  template_no: number;
  template_code: string | null;
  template_title: string | null;
  template_msg: string | null;
  button_name: string | null;
  button_link: string | null;
  button_link_name: string | null;
  button_mo: string | null;
  button_pc: string | null;
}